requirements.txt
and==0.1.1
beautifulsoup4==4.9.3
build==0.1.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Cython==0.29.21
gensim==3.8.3
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
install==1.3.4
MouseInfo==0.1.3
numpy==1.19.4
opencv-python==4.4.0.46
packaging==20.4
pandas==1.1.4
pep517==0.9.1
pgi==0.0.11.2
Pillow==8.0.1
pkg-resources==0.0.0
playsound==1.2.2
PyAudio==0.2.11
PyAutoGUI==0.9.52
pycairo==1.20.0
PyGetWindow==0.0.9
Pygments==2.7.2
PyGObject==3.38.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.6
python-dateutil==2.8.1
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.4
requests==2.24.0
scipy==1.5.4
selenium==3.141.0
six==1.15.0
smart-open==3.0.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
tesseract-ocr==0.0.1
to==0.3
toml==0.10.2
urllib3==1.25.11
